double Funtions Pract_4

+++ Practica_4/algorIte.h
@@ -1,4 +1,4 @@
-int factorialIte(int n);
+double factorialIte(int n);
 void fibonacci(int n);
 void torresDeHanoi(int n);
 void moverDisco(char desde, char hacia);
